Add files via upload
Week 1 - Exercise 3 : 
a.	Write a C# program to compute the sum of the first 500 prime numbers.
b.	Write a C# program to reverse the words of a sentence.
c.	Write a C# program to check if an array contains an odd number.
d.	Write a C# Sharp program that takes a decimal number as input and displays its equivalent in binary form.
e.	Write a program in C# Sharp to read any day number in integer and display day name in the word.
f.	Write a C# Sharp program to calculate the factorial of a given number.
g.	Write a program in C# Sharp to print all unique elements in an array.
h.	Write a program in C# Sharp to find the sum of first n natural numbers using recursion.